so mt burro

+++ src/B-tree/seek2.cpp
@@ -3,7 +3,7 @@
 #include <iostream>
 #include "../definitions.h"
 #include "secondary.h"
-
+// "Poster: 3D sketching and flexible input for surface design: A case study.";"2013";"Anamary Leal|Doug A. Bowman"
 int main(int argc, char const* argv[]) {
 
   cout << argv[1] << endl;
@@ -20,10 +20,10 @@ int main(int argc, char const* argv[]) {
   int numBlocosLidos;
   int totalBlocos;
   ArvoreBSec<char[300]> arvore(2, SECONDARY_INDEX_PATH);
-  unsigned long long int out = arvore.busca(titulo);
+  unsigned int out = arvore.busca(titulo);
   cout << "resultado: " << out << endl;
   fin.seekg(out);
-  
+
   string word;
   string wordOut;
   stringstream s1(word);
@@ -33,7 +33,7 @@ int main(int argc, char const* argv[]) {
   //   getline(fin, word);
   //   wordOut = word;
   //   // cout << "word " << word << endl;
-  //   stringstream s1(word);  
+  //   stringstream s1(word);
   //   std::getline(std::getline(s1, word, '"'), word, '"');
   // } while (stoi(word) != id);
